+++ src/Pipeline/FireAndForget.php
@@ -12,6 +12,7 @@
 
 namespace Predis\Pipeline;
 
+use Predis\Connection\AggregateConnectionInterface;
 use Predis\Connection\ConnectionInterface;
 use SplQueue;
 
@@ -25,13 +26,12 @@ class FireAndForget extends Pipeline
      */
     protected function executePipeline(ConnectionInterface $connection, SplQueue $commands)
     {
-        $buffer = '';
-
-        while (!$commands->isEmpty()) {
-            $buffer .= $commands->dequeue()->serializeCommand();
+        if ($connection instanceof AggregateConnectionInterface) {
+            $this->writeToMultiNode($connection, $commands);
+        } else {
+            $this->writeToSingleNode($connection, $commands);
         }
 
-        $connection->write($buffer);
         $connection->disconnect();
 
         return [];