+++ src/Pipeline/Pipeline.php
@@ -18,6 +18,7 @@ use Predis\ClientContextInterface;
 use Predis\ClientException;
 use Predis\ClientInterface;
 use Predis\Command\CommandInterface;
+use Predis\Connection\AggregateConnectionInterface;
 use Predis\Connection\ConnectionInterface;
 use Predis\Connection\Replication\ReplicationInterface;
 use Predis\Response\ErrorInterface as ErrorResponseInterface;
@@ -131,13 +132,11 @@ class Pipeline implements ClientContextInterface
      */
     protected function executePipeline(ConnectionInterface $connection, SplQueue $commands)
     {
-        $buffer = '';
-
-        foreach ($commands as $command) {
-            $buffer .= $command->serializeCommand();
+        if ($connection instanceof AggregateConnectionInterface) {
+            $this->writeToMultiNode($connection, $commands);
+        } else {
+            $this->writeToSingleNode($connection, $commands);
         }
-
-        $connection->write($buffer);
 
         $responses = [];
         $exceptions = $this->throwServerExceptions();
@@ -161,6 +160,39 @@ class Pipeline implements ClientContextInterface
         }
 
         return $responses;
+    }
+
+    /**
+     * Writes pipelined commands to single node connection.
+     *
+     * @param  ConnectionInterface $connection
+     * @param  SplQueue            $commands
+     * @return void
+     */
+    protected function writeToSingleNode(ConnectionInterface $connection, SplQueue $commands)
+    {
+        $buffer = '';
+
+        foreach ($commands as $command) {
+            $buffer .= $command->serializeCommand();
+        }
+
+        $connection->write($buffer);
+    }
+
+    /**
+     * Writes pipelined commands to multi node connection.
+     *
+     * @param  AggregateConnectionInterface $connection
+     * @param  SplQueue                     $commands
+     * @return void
+     */
+    protected function writeToMultiNode(AggregateConnectionInterface $connection, SplQueue $commands)
+    {
+        foreach ($commands as $command) {
+            $nodeConnection = $connection->getConnectionByCommand($command);
+            $nodeConnection->write($command->serializeCommand());
+        }
     }
 
     /**