+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 ### Added
+- Added cluster support for `XADD`, `XDEL` and `XRANGE` (#1587)
 ### Changed
+- Refactor pipeline data writing depends on connection type (#1586)
+
 ### Maintenance
 
 ## v3.2.0 (2025-08-05)